deleted checkbox from currency item layout and all code that related

+++ app/src/main/java/com/example/currencyconverter/CurrencyItem.kt
@@ -1,7 +1,7 @@
 package com.example.currencyconverter
 
 data class CurrencyItem(
-    var currencyId: Long = 0L,
+    var currencyId: Int = 0,
     var name: String = "",
     var image: Int = R.drawable.kazakhstan_flag,
     var value: Long = 0L,
